+++ examples/usb/host/usb_rndis_4g_module/components/espressif__usb_host_cdc_acm/host_test/main/test_parsing_modems.cpp
@@ -8,7 +8,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include "usb/usb_helpers.h"
 
-#include "cdc_descriptors.hpp"
+#include "descriptors/cdc_descriptors.hpp"
 #include "test_parsing_checker.hpp"
 #include "cdc_host_descriptor_parsing.h"
 #include "usb/cdc_acm_host.h"
@@ -191,6 +191,65 @@ SCENARIO("Modems descriptor parsing: 7000E", "[modem][7000E]")
         }
 
         // Interface 3-5 contain notification element
+        for (int interface = 3; interface <= 5; ++interface) {
+            SECTION("Interface " + std::to_string(interface)) {
+                cdc_parsed_info_t parsed_result = {};
+                esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, interface, &parsed_result);
+                REQUIRE_CDC_NONCOMPLIANT_WITH_NOTIFICATION(ret, parsed_result);
+            }
+        }
+    }
+}
+
+SCENARIO("Modems descriptor parsing: A7672E", "[modem][A7672E]")
+{
+    GIVEN("SimCom A7672E FS") {
+        const usb_device_desc_t *dev_desc = (const usb_device_desc_t *)sima7672e_device_desc_fs_hs;
+        const usb_config_desc_t *cfg_desc = (const usb_config_desc_t *)sima7672e_config_desc_fs;
+
+        // Interface 0-1 belongs USB_CLASS_WIRELESS_CONTROLLER class, expect no CDC interface
+        SECTION("Interface 0") {
+            cdc_parsed_info_t parsed_result = {};
+            esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, 0, &parsed_result);
+            REQUIRE(ret == ESP_ERR_NOT_FOUND);
+        }
+
+        // Diagnostic interface
+        SECTION("Interface 2") {
+            cdc_parsed_info_t parsed_result = {};
+            esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, 2, &parsed_result);
+            REQUIRE_CDC_NONCOMPLIANT(ret, parsed_result);
+        }
+
+        // Interface 3: NMEA, Interface 4-5: AT
+        for (int interface = 3; interface <= 5; ++interface) {
+            SECTION("Interface " + std::to_string(interface)) {
+                cdc_parsed_info_t parsed_result = {};
+                esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, interface, &parsed_result);
+                REQUIRE_CDC_NONCOMPLIANT_WITH_NOTIFICATION(ret, parsed_result);
+            }
+        }
+    }
+
+    GIVEN("SimCom A7672E HS") {
+        const usb_device_desc_t *dev_desc = (const usb_device_desc_t *)sima7672e_device_desc_fs_hs;
+        const usb_config_desc_t *cfg_desc = (const usb_config_desc_t *)sima7672e_config_desc_hs;
+
+        // Interface 0-1 belongs USB_CLASS_WIRELESS_CONTROLLER class, expect no CDC interface
+        SECTION("Interface 0") {
+            cdc_parsed_info_t parsed_result = {};
+            esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, 0, &parsed_result);
+            REQUIRE(ret == ESP_ERR_NOT_FOUND);
+        }
+
+        // Diagnostic interface
+        SECTION("Interface 2") {
+            cdc_parsed_info_t parsed_result = {};
+            esp_err_t ret = cdc_parse_interface_descriptor(dev_desc, cfg_desc, 2, &parsed_result);
+            REQUIRE_CDC_NONCOMPLIANT(ret, parsed_result);
+        }
+
+        // Interface 3: NMEA, Interface 4-5: AT
         for (int interface = 3; interface <= 5; ++interface) {
             SECTION("Interface " + std::to_string(interface)) {
                 cdc_parsed_info_t parsed_result = {};